Move fireMotionSensor to separate file
